requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
et_xmlfile==2.0.0
Flask==3.1.3
fonttools==4.62.1
fpdf2==2.8.7
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.19.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pillow==12.2.0
plotly==6.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.33.1
six==1.17.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
